distsqlrun: move disk_map.go to engine
This functionality is generally useful, and does not need to live in
distsqlrun.
